qcacld-3.0: Use atomic bit ops in ol_tx_desc_dup_detect functions
Bitmasks should use atomic operations when they can be
modified in multiple contexts.  Also sanitize msdu_id length
to avoid potential memory corruption.

CRs-Fixed: 1061195

+++ core/dp/txrx/ol_txrx_types.h
@@ -669,7 +669,7 @@ struct ol_txrx_pdev_t {
 		uint32_t offset_filter;
 		struct qdf_mem_multi_page_t desc_pages;
 #ifdef DESC_DUP_DETECT_DEBUG
-		uint32_t *free_list_bitmap;
+		unsigned long *free_list_bitmap;
 #endif
 	} tx_desc;
